removed interfaces; seems happy with cycling import dependency
